requirements.txt
# requirements.txt
langchain==0.6.1
openai==1.3.0
python-dotenv==1.0.0